requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.8.3
psycopg[binary]>=3.1.0
numpy>=1.26.0
scikit-learn>=1.5.0